Add Marking Features

+++ flot-android-project/src/com/flotandroidchart/flot/options/SeriesBars.java
@@ -20,7 +20,7 @@ package com.flotandroidchart.flot.options;
 public class SeriesBars {
 	public Boolean show = false;
 	public int lineWidth = 2;
-	public int barWidth = 1;
+	public float barWidth = 1;
 	public Boolean fill = true;
 	public Object fillColor = null;
 	public String align = "left";
